Modularize populations & projectors
