+++ openfisca_core/populations/config.py
@@ -1,0 +1,2 @@
+ADD = 'add'
+DIVIDE = 'divide'
