+++ openfisca_core/projectors/entity_to_person_projector.py
@@ -1,0 +1,12 @@
+from openfisca_core import projectors
+
+
+class EntityToPersonProjector(projectors.Projector):
+    """For instance person.family."""
+
+    def __init__(self, entity, parent = None):
+        self.reference_entity = entity
+        self.parent = parent
+
+    def transform(self, result):
+        return self.reference_entity.project(result)
